Support multiline inputs with bags [#26]
Upgrade rustyline from 6.2 -> 8.0 and change the Validator to suppport
bags.

The Validator is responsible for looking at a line and determining if it
is `Incomplete` or not. The previous implementation checked for
incomplete lines using `([{`, now we also check for bags (`<<`).

This was a little involved because we now need our own Validator (can't
use the one out of the box) and we also needed more fancy stuff to
support multi-character structures.

The net effect of this change is if you type

    qldb> begin
    qldb(tx: 7tiNWPtCr0V586d6es7wGJ)> INSERT INTO Person <<[press ENTER]

then you will be able to continue on the next line.

Note that in #26, the example is `INSERT INTO Person[press ENTER]` which
will not work. To make that case work, we'd need to parse the PartiQL
and note that the statement was not complete. (This example *does* work
if the text is pasted in, however.)

I think the next reasonable step would be to support explicit multiline
input by using a delimiter such as `;`. That would allow for fine-grain
control of submission without getting into parsing.