Fix compile on Windows
Utf8Error is Unix-only. For now, there are no tests that actually
trigger this LOC anyways, so just make it panic.

The problem is our trait returns a ReadlineError which is used by the
shell to do things like interpret ^C and ^D. So we can't do something
like 1/ use `anyhow::Result` or 2/ make Err an associated type on Ui,
because the shell needs to know *something* about the type of Error.

This can be fixed, but we'll need another layer of abstraction. For now,
I just wanted to fix the build.