New --terminator-required flag
If this is set to true (by default it is false), then pressing return
will NOT submit the input. Instead, the user is allowed to continuing
entering input until they end with the delimiter (currently set to a
semicolon).

For example, the following is now possible:

    qldbshell ... --terminator-required
    qldb> begin;
    qldb> insert into
    people {'name':'Bobby'}
    ;

This can be helpful when writing statements that are long!