Refactor to use the async driver
Before this commit, we were wrapping the underlying QldbDriver with a
blocking API. So, where the real driver has this API:

    async fn transact(...)

the blocking driver's API did not have 'async'. This made sense to me in
the early versions of the shell because the shell itself is NOT
asynchronous - it waits for humans!

However, the transact API takes a lambda argument to encapsulate good
use of the QLDB API such as 1/ idempotency, 2/ retries, 3/ calling
commit. This meant we needed to have code that looked like this:

```
driver.transact(|mut tx| {
    loop {
        // get and process user input
    }
})
```

So, because we want the user to be able to input many things, we needed
the IO loop inside the transaction lambda.

In recent commits, we refactored the "idle" mode to separate out the UI
loop and a particular tick of that loop. That was super useful for
testing purposes: we could set fake input, tick the loop, then see what
it did. That refactoring allowed us to write a nice test for "help", but
didn't help us at all testing driver interaction (like typing
abort/commit).

In this commit I completely redid the way this all works. The
transaction is now an "actor coroutine" with two sets of channels for
IO:

```
driver.transact(|mut tx| async {
    loop {
        // get and process user input..
        //   .. input comes from a Receiver NOT Stdin
        //   .. output goes out a Sender NOT Stdout
    }
})
```

With this change, the facilities we had for step-ticking the REPL loop
now just work. When a transaction is opened (with `begin`), a task is
spawned to handle that transaction. When a user enters PartiQL, that
PartiQL flows into the coroutine via its input channel, and statement
results come out the output channel.

I think this is going to let us remove the entire concept of modes, the
Option<Deps> hack and consolidate handling of commands. We remove the
entire notion of the BlockingQldbDriver here too!

There are two minor usability regressions as of this commit:

1. The transaction id is no longer displayed. I'm actually OK with
this (I thought it was too long and had dubious use, e.g. maybe logging
or something is a better way to show it).

2. You now need to type \commit or \abort. This is easy to fix, but I
wonder if we should (and, maybe the start/begin commands should also
require the command prefix).

I haven't fully cleaned up the code or added all the tests we can now
write. Hopefully future commits show off what a win this is!