Support autocommit
With this change, the shell will now interpret PartiQL *not* in a
transaction to be its own transaction. That is:

    qldb> select * from information_schema.user_tables

is the same as

    qldb> \begin
    qldb *> select * from information_schema.user_tables
    qldb *> \commit

This creates parity with the Python shell and is a direct customer ask,
because it helps with transaction timeouts.

This commit also bugfixes ^D when in a transaction (to call abort).