Add config, environment and settings
Up until now, the only way to configure the shell has been through the
CLI arguments, such as --auto-commit. The shell was coded against the
CLI arguments directly. This commit is the first step in fixing this.

The idea is that the shell will be coded against a mutable environment.
The environment starts out with defaults, is altered by configuration
and then further altered by CLI arguments.

In this commit, we add the environment, config and CLI mechanisms but
don't yet change the authority.

The actual implementation is a bit messy, but I think this commit
carries enough merit that I want to get it in.