Add a regression test for empty config files
See the previous commit for a reason why!

Also, no longer create an empty config file by default. Initially I
thought this was a good idea so that if users wanted to customize the
file the name would be obvious (since the file existed). However, the
name isn't obvious because you need to know the directory anyways (e.g.
I _just_ had to look this up on my Windows machine). So it feels better
to not make empty files for no real reason. There are probably other
ways of self-describing config (e.g. render the location in --help).