Implement --format table
This commit uses the comfy-table crate to render a "nested table" based
on the newly landed ion-rs Element API.

Elements are tuples of `(symbols, value)`, so for example:

```
person::{name:"Fred"}
```

the above Element has a single text annotation (person) and value (of
type struct). The value in turn has 1 field which itself is an element.
In this case, the element behind 'name' has no annotations and the value
type is a string with value "Fred"!

This abstraction is useful for us because it provides **random access**.
When building a table, like

```
+-----------+
| name      |
+===========+
| Elizabeth |
|-----------|
| Fred      |
+-----------+
```

it's important to know 1/ what the headers are and 2/ what the column
widths should be.

Of course, in our example, we have two documents with a single field.
But, it *may* be that different documents have fields in different
orders or different fields entirely.

The Element API makes it super convenient to collect the full set of
headers (across the result set) and then, as we iterate row by row, do
random access over the fields.

The comfy-table library takes care of figuring out row widths,
accounting for unicode width, and so on. There may be cross-platform (or
cross-terminal) rendering bugs - I don't know!

I've chosen to render nested content as nested tables:

```
qldb> select * from information_schema.user_tables
+--------+---------------------------------------------+------------------------+------+
| status | indexes                                     | tableId                | name |
+======================================================================================+
| ACTIVE |                                             | 7NmEOVTm9XCClLRhkuAzw0 | bar  |
|--------+---------------------------------------------+------------------------+------|
| ACTIVE | +-------+--------+------------------------+ | 1nmeTZw7uhuI2kBXrbYh8x | foo  |
|        | | expr  | status | indexId                | |                        |      |
|        | +=========================================+ |                        |      |
|        | | [bar] | ONLINE | 4zJJ3hSDASCJVv4PmuJXmH | |                        |      |
|        | +-------+--------+------------------------+ |                        |      |
+--------+---------------------------------------------+------------------------+------+
```

In the future, we may expose additional table layout options via
configuration flags. This may include options such as:

- the look and feel (comfy-table has themes, UTF8 rounded corners, etc.)
- max rows, max width, etc.
- pagination (press SPACE for next) to avoid reading everything into memory

Fixes #49