Refactor program and tick flow
Before this commit, the shell would make a single runner and then tell
it to do its thing. The runner would run until a `tick` returned false.

We want to support changing which ledger the shell is using. This really
comes down to which *driver* the shell is using (i.e.: endpoint, region,
ledger name, credentials).

A Runner has Deps and Deps has a driver. We *could* make Deps support
replacing the driver under the hood, but that would make all usages of
the driver a little unergonimic (it would require mutexes etc.). It
seems to me that changing drivers is _unusual_ and thus the code should
reflect that.

Thus, I think we should reinstantiate the Deps and Runner when we change
ledgers. To facilitate that, this commit starts putting in richer (than
a bool) flow control. The commented-out `loop` in `lib.rs` shows where I
intend to reinstantiate things.

The reason the loop is commented out is.. the compiler rejects the code.
Why? Because `Environment` is moved into `Runner::new_with_env`. This is
actually super exciting and, IMO, reinforces that this change is the
right one. We *want* the environment to persist across connection
changes (since it may be mutated, and we want those mutations to
persist).