Support forcing a newline
ALT+ENTER will force a newline regardless of 'terminator required' or
input completion. On Windows, the seq is ESC+ENTER to avoid conflicting
with maximize window.

Note that Py shell uses OPTION+ENTER, so this is a small change.
However, there is no good reason to vary between Linux/Mac and the web
shell is also currently ALT+ENTER.