Implement `set input-mode emacs`
Before this commit, all backslash commands (like `\help`) could only be
exact string matches. We need a path forward to accept args, and have
decent help.

This commit takes the stance that we can use structopt (which we already
use for CLI parsing) to also parse backslash commands. The first one
I've implemented is for changing the input mode from emacs to vi.

The Ui/readline infrastructure has been migrated to use Environment and
is able to respond directly to this new command.