Also disable query metrics in !tty mode
Ultimately we want to support

```
select * from example | qldb --format json | jq ...
```

and the query metrics will get in the way of that.

This commit is not really sufficient on its own (for example, if query
metrics are enabled in config, this commit will not disable them).
However, since you can't (yet) do that in config.. :)

We might want to offer different stats when in echo mode that are
printed to stderr.