Use ui.prompt (from config) to format the prompt
This commit will format the prompt according to the value in
`ui.prompt` (in the config file, if present). For example:

```
[ui]
prompt = "$REGION:$LEDGER$ACTIVE_TRANSACTION> "
```

```
us-west-2:shell-dev> begin
us-west-2:shell-dev *> select * from foo
```

Internally, environment was refactored to hold region as a value rather
than as an optional string. We didn't get that quite right (see the
FIXME that was added).