Reuse the healthcheck client
In a previous commit I made it way easier to trace the underlying
workings of the shell. I found that the first command had to get a new
connection - weird!

Obviously, this is because we had a new client for the healthcheck and
driver. So the healthcheck made a request, then we shut down that pool,
then made another. Now, we share the client!

This is especially nice for profile based clients because credentials
are now shared between clients too!

I also made a change to cleanup the healthcheck session.