Squash up config
The idea behind Setting and Setters was a nice one, but the
implementation is tricky. Having the CLI, config file and environment
each have to know about the set of settings (each in slightly different
ways) was making it hard to configure stuff. Ultimately I want to get
back to "traceable configuration", but I think the direction we had was
not the right one.

The new direction goes all in on the "config" idea, now known as
ShellConfig. It is defined using serde (and serde'd using toml). The
serde struct and annotations define what must be specified, what is
defaulted, etc.

The basic structure of ShellConfig is now:

- a set of configured ledgers (which might be empty)
- the default ledger to use if one isn't specified
- UI settings (such as whether autocommit is enabled)
- debug settings

The CLI only lets you change **ledger** settings. It can be used to pick
a ledger from the set of configured ledgers like so:

```toml
[[ledgers]]
name = "my-test-ledger"
region = "us-west-2"
```

```sh
$ qldb --ledger my-test-ledger
```

This means that CLI retains the ability to connect to a ledger without
any shell config on disk at all, or provide zero arguments and rely
entirely on shell config:

```toml
default_ledger = "my-test-ledger"
```

This pattern makes unix piping a pleasure:

```
cat query.partiql | qldb
```

In the future, we will allow for other settings (UI, debug) to be
overridden on a per-ledger basis or via the CLI.

The environment now "has a" config instead of the previous "is a"
relationship. Any config value can be read via:

```rust
let _ = env.config().debug.log;
```

.. whereas previously we had to plumb each value to the top level (and
it didn't reflect the tree structure):

```rust
let _ = env.debug_log_file().value;
```

As a side-effect of this change, some CLI args were
removed (``--terminator-required`, `--auto-commit` and `--execute`).
I've already explained the rationale for the first two. The last one was
just not very well thought out and made irrelevant by making unix-style
piping work better.