Bugfix ping, also broken by the config refactor
This was caused by a deadlock when building the endpoint (the env lock
on current_ledger() was held and then the same lock for current_region()
was queued).

To fix this, I moved from Mutex to RwLock. While I *could* have
refactored the code to avoid the double lock acquire, I think that
RwLock is just a better fit for what we need.