Fix a bug with expired transactions
As reported in #79, an expired transaction would cause a panic to be
printed on the next command:

```
qldb> begin
// do nothing for a while
qldb *> select * from information_schema.user_tables
// transaction has expired warning (expected)
qldb> begin
qldb *> select * from information_schema.user_tables
// previous transaction panics, crashes a tokio thread
// lots of stack trace
// results from transaction 2 (which worked just fine)
```

The reason for this is the first transaction was sending an `abort` on a
session that was marked invalid. It was sending an abort because
cancellation wasn't setup properly. In particular, after the first
transaction is dropped, it was cancelled "slowly", allowing it to finish
any other work. A sentinel value went into its queue, causing an attempt
to abort the (already expired) transaction. This in turned triggered an
assertion.

With this fix, the future is dropped synchronously. The sentinel never
goes down the channel, and the transaction abort attempt never happens.

The reason for the panic is that the expired transaction returns an
InvalidSessionException which marks the pool connection as invalid. So,
when the abort came through, that triggered an assertion that this
connection should no longer be used. That is a bug that should be fixed
in the driver (it should return an illegal state, not a panic).

Fixes #79.