Fail to move to the awssdk driver
Credentials fail to load (they hang) due to a future not running. I
suspect I either need to figure out how to use credentials from the
awssdk directly OR give up on the current thread runtime.

The core issue here is that the rusoto credentials are asynchronous
while the new ones aren't. Briding this is hard.

This commit completely breaks the shell (it hangs).