Fix credential bridge
I really didn't want to do this, but it seems the only way is to run
another runtime on another thread. So that's what we do now.

I _could_ have changed the main runtime to be multi-thread and then I
could use block_on directly, but, I didn't want to make so broad a
change for so local a concern.

I'll get rid of the rusoto credential bridge soon. In other news, region
defaulting is broken.