Make get job ID more robust
Originally we were sleeping for 5 seconds to get object tag which contains the job ID. If it was still not there, then the job would fail. 

To make it more robust, I added a for loop to retry up to 3 times and wait 5 seconds between retries. If it still can't retrieve the job ID by the end, then the job would fail with a message.