BadRequestException no longer breaks a transaction
I initially had this behavior because it mirrors the default experience
when programming against Qldb (exceptions throw, causing the transaction
to fail). But, of course, a programmer might catch the exception and
continue. This meant the shell wasn't allowing developers to actually
understand how the service worked.

Also, minor typos (e.g. of table names) would fail the transaction. More
likely, the user could use UP to repeat the query with a fix and it'd
all work out just fine.