default_config.toml -> config.ion
This commit changes the format of the config file from TOML to
Ion (text) as discussed in #141. This reinforces the concept of Ion with
customers, and is pretty straightforward because of how Ion text is a
superset of JSON.

The name of the file is also changed, and the documentation is updated
to reflect the new defaults.

Ion-rust doesn't yet have serde bindings, so I had to write the deser
out by hand.