Abort autocommit statements with errors
Before this commit, a single statement transaction (in autocommit mode)
might result in a hang, as reported in #142.

The reason for this is that not all errors end the transaction on the
server. So, for example, `select * from foo` might throw an error if
there is no table "foo", but the *transaction is still open*.

This means that the session goes back into the pool and the next
transaction attempt fails, by calling StartTransaction where one was
already open on that session.

There should be defense in depth in the pool implementation to prevent
the hang. However, this is the simplest change to fix the bug.