fix case where no credentials are found
The rusoto credentials bridge was added before the awssdk supported
async providers. It had a bug, causing a bad user message as
demonstrated in #166. To fix the bug, we move to the new async interface
which also cleans up some code.

Note that the overall experience is still not wonderful because the
default chain tries to connect to localhost (to support instance
metadata credentials). This has a 30s timeout. To fix this, we'll need
to either workaround the behavior in rusoto or wait for the awssdk to
provide in-crate providers.